Mod Base v1.5.0
This file is critical!!! If you accidentally deleted it thinking it was a mod, and now no compiled mod is loading, this is why. The mod manager needs this file, it includes all the base code used to execute API calls from your mods. This Mod Base is designed to work with ACMM v1.5.0 specifically, using the one from v1.0.0 will NOT work.